Update 02_headers.md
Hello, this is fun. I hope it is received well. I've taken time to add my $0.02 worth to improve this lesson. I was confused at this point in the lesson and had to guess the next step of **...** to edit the file. A newbie could easily be lost at this point and lose interest and faith in the lesson. My changes are revealed by comparison and are not repeated here. I am not alone in the confusion created by this existing lesson; let us change that!  Thanks.